refactor: Change package name from core to adk for consistency across agent files
