+++ examples/docker-compose/a2a/helloworld-agent/adk/agent.go
@@ -1,4 +1,4 @@
-package core
+package adk
 
 import (
 	"context"
