+++ examples/docker-compose/a2a/helloworld-agent/adk/toolshandler.go
@@ -1,4 +1,4 @@
-package core
+package adk
 
 import (
 	sdk "github.com/inference-gateway/sdk"
